requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
lxml==3.7.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
requests==2.13.0
schedule==0.4.2
six==1.10.0
